Fix allowance fetch error
