+++ package.json
@@ -14,7 +14,7 @@
     "@chakra-ui/react": "^2.4.6",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
-    "@railgun-community/quickstart": "^3.7.2",
+    "@railgun-community/quickstart": "^3.8.3",
     "@railgun-community/shared-models": "^3.6.3",
     "@rainbow-me/rainbowkit": "^0.8.1",
     "browser-level": "^1.0.1",
