+++ pages/_app.tsx
@@ -1,5 +1,6 @@
 import Header from "@/components/Header";
 import { TokenListProvider } from "@/contexts/TokenContext";
+import { loadProviders } from "@/utils/railgun";
 import { initialize } from "@/utils/railgun";
 import { Grid, GridItem } from "@chakra-ui/layout";
 import { ChakraProvider } from "@chakra-ui/react";
@@ -8,7 +9,7 @@ import { RainbowKitProvider, getDefaultWallets } from "@rainbow-me/rainbowkit";
 import "@rainbow-me/rainbowkit/styles.css";
 import type { AppProps } from "next/app";
 import Head from "next/head";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { configureChains, createClient, WagmiConfig } from "wagmi";
 import { mainnet, polygon, arbitrum, bsc, goerli } from "wagmi/chains";
 import { alchemyProvider } from "wagmi/providers/alchemy";
@@ -47,6 +48,15 @@ const wagmiClient = createClient({
 
 function MyApp({ Component, pageProps }: AppProps) {
   initialize();
+  const [isProviderLoaded, setProviderLoaded] = useState<Boolean>(false);
+  useEffect(() => {
+    const fn = async () => {
+      await loadProviders();
+      setProviderLoaded(true);
+    };
+    fn();
+  }, []);
+
   return (
     <>
       <Head>
@@ -60,23 +70,25 @@ function MyApp({ Component, pageProps }: AppProps) {
       <WagmiConfig client={wagmiClient}>
         <RainbowKitProvider chains={chains}>
           <ChakraProvider theme={theme}>
-            <TokenListProvider>
-              <Grid
-                templateAreas={`". header ."
+            {isProviderLoaded && (
+              <TokenListProvider>
+                <Grid
+                  templateAreas={`". header ."
                   ". body ."`}
-                gridTemplateRows={"4.8rem 100vh"}
-                gridTemplateColumns={"1fr minmax(auto, 150rem) 1fr"}
-                gap="1"
-                marginX="1rem"
-              >
-                <GridItem area={"header"}>
-                  <Header />
-                </GridItem>
-                <GridItem area={"body"}>
-                  <Component {...pageProps} />
-                </GridItem>
-              </Grid>
-            </TokenListProvider>
+                  gridTemplateRows={"4.8rem 100vh"}
+                  gridTemplateColumns={"1fr minmax(auto, 150rem) 1fr"}
+                  gap="1"
+                  marginX="1rem"
+                >
+                  <GridItem area={"header"}>
+                    <Header />
+                  </GridItem>
+                  <GridItem area={"body"}>
+                    <Component {...pageProps} />
+                  </GridItem>
+                </Grid>
+              </TokenListProvider>
+            )}
           </ChakraProvider>
         </RainbowKitProvider>
       </WagmiConfig>
