+++ pages/index.tsx
@@ -7,20 +7,12 @@ import Head from "next/head";
 import { useEffect, useState } from "react";
 
 const Home: NextPage = () => {
-  const [isProviderLoaded, setProviderLoaded] = useState<Boolean>(false);
-  useEffect(() => {
-    const fn = async () => {
-      await loadProviders();
-      setProviderLoaded(true);
-    };
-    fn();
-  }, []);
   return (
     <Flex direction="column" align="center" justify="center" pt="2rem">
       <Heading as="h1" size="4xl" mb="4rem">
         Shield
       </Heading>
-      {isProviderLoaded && <TxForm />}
+      <TxForm />
       <HowTo />
     </Flex>
   );
